Multiorder maps (#39)
With this pull request I would like to include in skyreader an
implementation to save scans as multiorder maps.
The idea is to save exactly the same pixels as the ones which were
scanned, and for each report the density of probability (or density of
llh [deg-2] if a llh map is desired).

To do this, I needed to add some functions in `handle_map_data.py`, as
for some directions there are multiple pixels scanned with different
nsides. In these cases, the pixels with the bigger nside need to be
ignored.

Moreover, I added some logic in `extract_map` to fill the bigger nside
with empty pixels. This is necessary to produce multiorder maps for
pointed scans as well. It does not cause problems in the rest of the
processing as further on the empty pixels are filled with the smallest
registered probability.

Happy to receive feedback about this!

---------

Co-authored-by: wipacdevbot <developers@icecube.wisc.edu>