Compatibility with lvk (#41)
To be more compatible with LVK, I:
- Changed the unit in the multiorder maps from deg-2 to sr-1;
- The unit is stored now in `TUNIT2` in the header;
- Changed the column name from `PROBABILITY DENSITY [deg-2]` to
`[PROBDENSITY]`.

Moreover, in `plot.py` there were problems bacause the new matplotlib
doesn't support anymore the attribute `QuadContourSet.collections`. This
affects only `create_plot` when `dozoom` is set to `True`. I tried by
using `get_paths()` to fix the problem, but I am not totally sure it is
really doing what it should. At least in general it is not anymore
giving error. However, is the option `dozoom=True` still used? It seems
to me only redundant because there is already `create_plot_zoomed`.
Could we get rid of the option `dozoom=True`?

---------

Co-authored-by: wipacdevbot <developers@icecube.wisc.edu>