Add information to the header of the reco fits file (rectangular errors at 50% + contour areas at 50% and 90%) (#50)
Hi,

I propose a slight modification to the function that generates zoomed
plots from reconstructions.
The modification would allow it to save RA and Dec, as well as the
corresponding 50% and 90% rectangular and circular areas.

https://github.com/icecube/skyreader/blob/40a75068b486225fa02c433caf6cdd76720ddf1e/skyreader/plot/plot.py#L729

This change would make it easier to manage reconstruction outputs in
general, and is necessary for posting reconstruction results on Slack
channels as part of the automation I am currently working on.

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Angela <angela@mobile-222.shadow2.airub.net>
Co-authored-by: ric-evans <emejqz@gmail.com>