[minor] Fix pixel ordering and selection of best-fit direction (#75)
The switch to a probability map is applied (with an eventual Gaussian
smoothing) to the `equatorial_map`. This is subsequently applied to
`grid_value` (the actual multiorder map) through an interpolation. Both
the Gaussian smoothing and the interpolation can slightly change the
position of the best-fit direction. Previously, the best-fit direction
was taken from `grid_value` before the Gaussian smoothing and the
interpolation. This created mismatches between the best-fit direction in
the header and the actual best-fit direction in the multiorder map. The
pull request aims to fix this by taking the best-fit direction after the
application of smoothing and interpolation.